requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
sqlparse==0.5.1
whitenoise==6.8.2
